requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.7.0
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
regex==2020.1.7
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
